# internal/api/api.go
// Package api implements the JSON management REST API: repositories, and (from
// Phase 3) users, roles, group mappings and personal access tokens.
package api

import (
	"context"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/younsl/o/box/kubernetes/forklift/internal/audit"
	"github.com/younsl/o/box/kubernetes/forklift/internal/auth"
	"github.com/younsl/o/box/kubernetes/forklift/internal/meta"
	"github.com/younsl/o/box/kubernetes/forklift/internal/notify"
	"github.com/younsl/o/box/kubernetes/forklift/internal/version"
)

// Handler serves the management API.
type Handler struct {
	store  *meta.Store
	authz  *auth.Service
	log    *slog.Logger
	client *http.Client
	rec    *audit.Recorder
	// haStatus assembles live HA/leadership status; injected from main once
	// leader election and the storage backend are known. Nil in tests.
	haStatus func(context.Context) HAStatus
	// haStepDown asks this instance to voluntarily release leadership for a
	// manual failover, reporting whether it was the leader (and thus stepped
	// down). Nil in single-instance mode and tests.
	haStepDown func() bool
	// notifier delivers outbound alarms; backs the receiver test, repository
	// sample-send and preview endpoints. Injected from main; nil in tests.
	notifier *notify.Notifier
}

// SetNotifier injects the outbound-alarm notifier used by the receiver test and
// repository sample/preview endpoints.
func (h *Handler) SetNotifier(n *notify.Notifier) { h.notifier = n }

// New creates an API handler. authz may be nil in tests that exercise only
// public endpoints and rec may be nil to disable audit logging, but production
// wiring always provides both.
func New(store *meta.Store, authz *auth.Service, log *slog.Logger, rec *audit.Recorder) *Handler {
	return &Handler{
		store:  store,
		authz:  authz,
		log:    log,
		client: &http.Client{Timeout: 5 * time.Second},
		rec:    rec,
	}
}

// audit records a repository lifecycle event performed through the management
// API, attributed to the authenticated principal.
func (h *Handler) audit(r *http.Request, repoName, event string, status int) {
	var username string
	if p := auth.FromContext(r.Context()); p != nil {
		username = p.Username
	}
	h.rec.Record(audit.Event{
		Repo:      repoName,
		Action:    event,
		Username:  username,
		Method:    r.Method,
		Status:    status,
		ClientIP:  audit.ClientIP(r),
		UserAgent: r.UserAgent(),
	})
}

// Routes returns a router mounted under /api/v1. The auth.Service middleware is
// applied by the caller, so a principal (if any) is already in the context.
func (h *Handler) Routes() chi.Router {
	r := chi.NewRouter()

	// Public.
	r.Post("/login", h.login)
	r.Post("/logout", h.logout)
	r.Get("/me", h.me)
	r.Get("/version", h.version)

	// Authenticated self-service: personal access tokens.
	r.Group(func(r chi.Router) {
		if h.authz != nil {
			r.Use(h.authz.RequireAuth)
		}
		r.Get("/tokens", h.listTokens)
		r.Post("/tokens", h.createToken)
		r.Delete("/tokens/{id}", h.deleteToken)
		// Names only, for token-scope autocomplete.
		r.Get("/repository-names", h.listRepositoryNames)
		// Repository listing and read-only detail (config + artifact browse) are
		// available to any authenticated user; the handlers filter to repositories
		// the principal can read. Mutations, audit logs and remote-health stay
		// admin-only (below). Mirrors Nexus browse/read vs admin privileges.
		r.Get("/repositories", h.listRepositories)
		r.Get("/repositories/{id}", h.getRepository)
		r.Get("/repositories/{id}/artifacts", h.listArtifacts)
		r.Get("/repositories/{id}/artifacts/scan", h.getArtifactScan)
		r.Get("/repositories/{id}/artifacts/sbom", h.getArtifactSBOM)
	})

	// Package approvals: administrators plus principals holding the approve
	// action (e.g. a security-engineer role). The queue is shared; decisions
	// are enforced per repository inside the handlers.
	r.Group(func(r chi.Router) {
		if h.authz != nil {
			r.Use(h.authz.RequireApprover)
		}
		r.Route("/approvals", func(r chi.Router) {
			r.Get("/", h.listApprovals)
			r.Get("/count", h.countApprovals)
			r.Get("/{id}", h.getApproval)
			r.Post("/", h.createApproval)
			r.Post("/approve-all", h.approveAllPending)
			r.Post("/{id}/approve", h.approveApproval)
			r.Post("/{id}/reject", h.rejectApproval)
		})
		r.Route("/version-denies", func(r chi.Router) {
			r.Get("/", h.listVersionDenies)
			r.Post("/", h.createVersionDeny)
			r.Delete("/{id}", h.deleteVersionDeny)
		})
	})

	// Administrative reads: administrators plus principals holding the audit
	// action (e.g. a security auditor). Read-only views of the admin surfaces.
	r.Group(func(r chi.Router) {
		if h.authz != nil {
			r.Use(h.authz.RequireAuditor)
		}
		r.Get("/repositories/{id}/upstream-health", h.upstreamHealth)
		r.Get("/repositories/{id}/audit-logs", h.listAuditLogs)
		r.Get("/repositories/{id}/permissions", h.repositoryPermissions)
		r.Get("/repositories/{id}/tokens", h.repositoryTokens)
		r.Get("/users", h.listUsers)
		r.Get("/users/{id}/tokens", h.listUserTokens)
		r.Get("/roles", h.listRoles)
		r.Get("/group-mappings", h.listGroupMappings)
		// Notification receivers are listed for admins/auditors so the repository
		// settings UI can offer them for selection.
		r.Get("/notification/receivers", h.listReceivers)
	})

	// Administrative mutations: administrators only. Same paths as the reads
	// above are registered here per-method, so chi applies the admin middleware
	// to the mutating verbs while reads stay under RequireAuditor.
	r.Group(func(r chi.Router) {
		if h.authz != nil {
			r.Use(h.authz.RequireAdmin)
		}
		// HA/leadership status for the management console (admin-only read).
		r.Get("/ha", h.getHAStatus)
		// Manual failover: ask the leader to step down so a standby takes over.
		r.Post("/ha/step-down", h.stepDownHA)
		r.Post("/repositories", h.createRepository)
		r.Post("/repositories/check-upstream", h.checkUpstream)
		r.Put("/repositories/{id}", h.updateRepository)
		r.Post("/repositories/{id}/disabled", h.setRepositoryDisabled)
		r.Post("/repositories/{id}/artifacts/scan", h.scanArtifact)
		r.Post("/repositories/{id}/artifacts/scan-batch", h.scanArtifactsBatch)
		r.Post("/artifact-scans/scan-all", h.scanAllArtifacts)
		r.Post("/artifact-scans/verdicts/recompute", h.recomputeArtifactScanVerdicts)
		r.Post("/artifact-scans/sboms/export", h.exportArtifactSBOM)
		r.Delete("/repositories/{id}", h.deleteRepository)
		r.Delete("/repositories/{id}/artifacts", h.deleteArtifact)
		r.Post("/users", h.createUser)
		r.Put("/users/{id}", h.updateUser)
		r.Delete("/users/{id}", h.deleteUser)
		r.Post("/users/{id}/roles", h.assignRole)
		r.Delete("/users/{id}/roles/{roleID}", h.removeRole)
		r.Post("/users/{id}/tokens", h.createUserToken)
		r.Delete("/users/{id}/tokens/{tokenID}", h.deleteUserToken)
		r.Post("/roles", h.createRole)
		r.Delete("/roles/{id}", h.deleteRole)
		r.Post("/roles/{id}/permissions", h.addPermission)
		r.Delete("/roles/{id}/permissions/{permID}", h.deletePermission)
		r.Post("/group-mappings", h.createGroupMapping)
		r.Delete("/group-mappings/{id}", h.deleteGroupMapping)
		r.Post("/notification/receivers", h.createReceiver)
		r.Put("/notification/receivers/{id}", h.updateReceiver)
		r.Delete("/notification/receivers/{id}", h.deleteReceiver)
		r.Post("/notification/receivers/{id}/test", h.testReceiver)
		r.Post("/notification/test", h.testWebhookAdhoc)
		r.Get("/repositories/{id}/notification/sample", h.previewRepoSample)
		r.Post("/repositories/{id}/notification/sample", h.sendRepoSample)
	})

	return r
}

// version reports the build-time version metadata so the web UI can show it in
// the sidebar. Public: it leaks nothing sensitive and aids support triage.
func (h *Handler) version(w http.ResponseWriter, _ *http.Request) {
	// oidc_enabled drives the login page: the "Sign in with Keycloak" button is
	// hidden when OIDC is not configured (its /auth/login route is unregistered).
	writeJSON(w, http.StatusOK, map[string]any{
		"version":      version.Version,
		"commit":       version.Commit,
		"oidc_enabled": h.authz != nil && h.authz.OIDCEnabled(),
	})
}

// validName accepts the identifier charset shared by every "name" input
// (repository, token, role, user names): ASCII letters, digits, '-' and '_',
// at most 64 characters. Descriptions and notes stay free-form.
func validName(name string) bool {
	if name == "" || len(name) > 64 {
		return false
	}
	for _, c := range name {
		switch {
		case c >= 'a' && c <= 'z', c >= 'A' && c <= 'Z', c >= '0' && c <= '9', c == '-', c == '_':
		default:
			return false
		}
	}
	return true
}

const nameRuleMsg = "may only contain letters, digits, '-' and '_' (max 64 chars)"

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json; charset=utf-8")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}

func writeError(w http.ResponseWriter, status int, msg string) {
	writeJSON(w, status, map[string]string{"error": msg})
}

// mapError translates store errors into HTTP responses.
func mapError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, meta.ErrNotFound):
		writeError(w, http.StatusNotFound, "not found")
	case errors.Is(err, meta.ErrConflict):
		writeError(w, http.StatusConflict, "conflict")
	default:
		writeError(w, http.StatusInternalServerError, err.Error())
	}
}