# internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: forklift API
  version: "0.1.0"
  description: >
    Management API for forklift, a Kubernetes-native artifact repository.
    Package-format endpoints (Maven, npm, Cargo, Go, PyPI) are mounted under
    /maven, /npm, /cargo, /go and /pypi and follow each ecosystem's native protocol.
servers:
  - url: /
...
paths:
  /api/v1/login:
    post:
      tags: [session]
      summary: Log in with a local username and password (sets a session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { description: Logged in }
        "401": { description: Invalid credentials }
...
  /api/v1/repositories:
    get:
      tags: [repositories]
      summary: List repositories
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Repository" }
    post:
      tags: [repositories]
      summary: Create a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepositoryCreate" }
      responses:
        "201": { description: Created }
        "400": { description: Validation error }
        "409": { description: Name already exists }
  /api/v1/repositories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [repositories]
      summary: Get a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Repository
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repository" }
        "404": { description: Not found }
    put:
      tags: [repositories]
      summary: Update a repository's upstream URL and config
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Updated }
    delete:
      tags: [repositories]
      summary: Delete a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
  /api/v1/repositories/{id}/audit-logs:
    get:
      tags: [repositories]
      summary: List a repository's audit log (admin), newest first
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - name: event
          in: query
          schema:
            type: string
            enum: [download, upload, delete, repo.create, repo.update, repo.delete,
                   approval.request, approval.approve, approval.reject]
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  logs:
                    type: array
                    items: { $ref: "#/components/schemas/AuditLog" }
        "404": { description: Not found }
  /api/v1/repository-names:
    get:
      tags: [repositories]
      summary: List repository names (any authenticated user) for token-scope autocomplete
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Slim repository list (name, format, type)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RepositoryName" }
  /api/v1/repositories/check-upstream:
    post:
      tags: [repositories]
      summary: Probe an upstream URL for reachability before creating a proxy (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, description: Absolute http(s) URL to probe. }
      responses:
        "200":
          description: Connectivity result (any HTTP response counts as reachable)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamHealth" }
        "400": { description: url is required }
  /api/v1/repositories/{id}/artifacts:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [repositories]
      summary: List a repository's artifacts (any user who can read the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: prefix, in: query, schema: { type: string }, description: Filter by artifact path prefix. }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Artifacts with aggregate count and total size
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactList" }
        "404": { description: Not found }
    delete:
      tags: [repositories]
      summary: Delete one artifact, or purge all artifacts in the repository (admin)
      description: >
        With a "path" query parameter, deletes that single artifact (204). Without it,
        purges every artifact in the repository and returns the number deleted (200).
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: path, in: query, schema: { type: string }, description: Artifact path to delete. Omit to purge all artifacts. }
      responses:
        "200":
          description: Purged all artifacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }
        "204": { description: Single artifact deleted }
        "404": { description: Not found }
  /api/v1/repositories/{id}/artifacts/scan:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
      - { name: path, in: query, required: true, schema: { type: string }, description: Artifact path to inspect or enqueue for scanning. }
    get:
      tags: [repositories]
      summary: Get the latest artifact-byte scan details for one artifact
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Latest scan status, package coordinate, and normalized findings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactScanResult" }
        "400": { description: path is required }
        "404": { description: Not found }
    post:
      tags: [repositories]
      summary: Queue a manual artifact-byte scan for one artifact
      description: >
        Enqueues work only. Scanner tools run in an external scanner worker
        that claims jobs through the internal scan API.
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "202":
          description: Scan job queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactScanJob" }
        "400": { description: path is required }
        "404": { description: Not found }
  /api/v1/repositories/{id}/artifacts/scan-batch:
    post:
      tags: [repositories]
      summary: Queue manual artifact-byte scans for multiple artifacts
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArtifactScanBatchRequest" }
      responses:
        "202":
          description: Scan jobs queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactScanBatchResponse" }
  /api/v1/repositories/{id}/artifacts/sbom:
    get:
      tags: [repositories]
      summary: Get the latest stored SBOM JSON for one artifact
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: path, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Stored SBOM document
          content:
            application/json:
              schema: { type: object }
        "404": { description: Not found }
  /api/v1/artifact-scans/scan-all:
    post:
      tags: [artifact-scans]
      summary: Queue artifact-byte scans for a repository or all repositories
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArtifactScanAllRequest" }
      responses:
        "202":
          description: Scan jobs queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactScanOpsResponse" }
  /api/v1/artifact-scans/verdicts/recompute:
    post:
      tags: [artifact-scans]
      summary: Recompute artifact scan verdicts without rescanning bytes
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArtifactScanRecomputeRequest" }
      responses:
        "200":
          description: Verdicts recomputed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactScanOpsResponse" }
  /api/v1/artifact-scans/sboms/export:
    post:
      tags: [artifact-scans]
      summary: Request export of a stored SBOM to an external destination
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArtifactSBOMExportRequest" }
      responses:
        "202":
          description: Export request recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactSBOMExport" }
        "400": { description: Invalid request body, no paths, or more than 100 paths. }
        "404": { description: Not found }
  /api/v1/repositories/{id}/disabled:
    post:
      tags: [repositories]
      summary: Take a repository offline or back online (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disabled]
              properties:
                disabled: { type: boolean }
      responses:
        "200":
          description: Updated repository
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repository" }
        "404": { description: Not found }
...
  /api/v1/repositories/{id}/permissions:
    get:
      tags: [repositories]
      summary: List role permissions granting access to this repository (admin/auditor)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Role permissions whose pattern matches this repository
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RepoPermission" }
        "404": { description: Not found }
  /api/v1/repositories/{id}/tokens:
    get:
      tags: [repositories]
      summary: List personal access tokens that can reach this repository (admin/auditor)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Tokens reaching this repository
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RepoToken" }
        "404": { description: Not found }
  /api/v1/tokens:
    get:
      tags: [tokens]
      summary: List the current user's personal access tokens
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Tokens }
    post:
      tags: [tokens]
      summary: Create a personal access token (plaintext returned once)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201": { description: Created }
  /api/v1/tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke a personal access token
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Revoked }
  /api/v1/users:
    get:
      tags: [users]
      summary: List users (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Users }
    post:
      tags: [users]
      summary: Create a local user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "201": { description: Created }
  /api/v1/users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      tags: [users]
      summary: Update a user (admin) — reset password and/or toggle disabled
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, description: New password (local users only). }
                disabled: { type: boolean }
      responses:
        "200": { description: Updated user }
        "400": { description: Validation error (incl. disabling your own account) }
        "404": { description: Not found }
    delete:
      tags: [users]
      summary: Delete a user (admin); deleting your own account is rejected
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "400": { description: Cannot delete your own account }
  /api/v1/users/{id}/roles:
    post:
      tags: [users]
      summary: Assign a role to a user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id: { type: integer }
      responses:
        "204": { description: Assigned }
  /api/v1/users/{id}/roles/{roleID}:
    delete:
      tags: [users]
      summary: Remove a role from a user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: roleID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
  /api/v1/users/{id}/tokens:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [tokens]
      summary: List a user's personal access tokens (admin or auditor)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Tokens }
        "404": { description: Not found }
    post:
      tags: [tokens]
      summary: Create a personal access token for a user (admin); plaintext returned once
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201": { description: Created }
        "404": { description: Not found }
  /api/v1/users/{id}/tokens/{tokenID}:
    delete:
      tags: [tokens]
      summary: Revoke a user's personal access token (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: tokenID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Revoked }
  /api/v1/roles:
    get:
      tags: [roles]
      summary: List roles with their permissions (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Roles }
    post:
      tags: [roles]
      summary: Create a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "201": { description: Created }
  /api/v1/roles/{id}:
    delete:
      tags: [roles]
      summary: Delete a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /api/v1/roles/{id}/permissions:
    post:
      tags: [roles]
      summary: Add a permission to a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "201": { description: Created }
  /api/v1/roles/{id}/permissions/{permID}:
    delete:
      tags: [roles]
      summary: Remove a permission from a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: permID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
  /api/v1/approvals:
    get:
      tags: [approvals]
      summary: List package approval requests and decisions (admin or approve permission), newest first
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: repo, in: query, schema: { type: string } }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Approval rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  approvals:
                    type: array
                    items: { $ref: "#/components/schemas/PackageApproval" }
    post:
      tags: [approvals]
      summary: Manually approve or reject a package (admin or approve permission on the repository), even before any request
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo, package, status]
              properties:
                repo: { type: string }
                package: { type: string, description: Canonical per-format package name. }
                status: { type: string, enum: [approved, rejected] }
                note: { type: string }
      responses:
        "201":
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "400": { description: Invalid request or non-proxy repository }
        "404": { description: Repository not found }
  /api/v1/approvals/count:
    get:
      tags: [approvals]
      summary: Count approval rows (admin or approve permission), for the pending badge
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: repo, in: query, schema: { type: string } }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        "200":
          description: Row count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
  /api/v1/approvals/approve-all:
    post:
      tags: [approvals]
      summary: Approve every pending package in one proxy repository (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo]
              properties:
                repo: { type: string }
                note: { type: string }
      responses:
        "200":
          description: Number of pending packages approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  approved: { type: integer }
        "400": { description: Missing repo or repo is not a proxy }
        "404": { description: Repository not found }
  /api/v1/approvals/{id}:
    get:
      tags: [approvals]
      summary: Get a package approval by id
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Package approval
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "404": { description: Not found }
  /api/v1/approvals/{id}/approve:
    post:
      tags: [approvals]
      summary: Approve a package (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "404": { description: Not found }
  /api/v1/approvals/{id}/reject:
    post:
      tags: [approvals]
      summary: Reject a package (admin or approve permission on the repository); cached content stops being served immediately
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "404": { description: Not found }
...
  /api/v1/version-denies/{id}:
    delete:
      tags: [approvals]
      summary: Remove a deny entry (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
        "404": { description: Not found }
...
  /api/v1/ha:
    get:
      tags: [ha]
      summary: High-availability/leadership status (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: HA status snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HAStatus" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Local username/password or a personal access token as the password.
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token (flpat_...) or an OIDC bearer token.
  schemas:
    Repository:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        format: { type: string, enum: [maven, npm, cargo, go, pypi] }
        type: { type: string, enum: [hosted, proxy, group] }
        upstream_url: { type: string }
        config: { $ref: "#/components/schemas/RepoConfig" }
        artifact_count:
          type: integer
          description: Stored artifact count (list endpoint only).
        total_size:
          type: integer
          description: Total stored bytes (list endpoint only).
    RepositoryCreate:
      type: object
      required: [name, format, type]
      properties:
        name: { type: string }
        format: { type: string, enum: [maven, npm, cargo, go, pypi] }
        type: { type: string, enum: [hosted, proxy, group] }
        upstream_url: { type: string }
        config: { $ref: "#/components/schemas/RepoConfig" }
...
    ArtifactList:
      type: object
      properties:
        count: { type: integer }
        total_size: { type: integer, format: int64 }
        artifacts:
          type: array
          items: { $ref: "#/components/schemas/Artifact" }
...